login_flask_app - MySQL & REST(FLASK-WTF)/requirements.txt
mysql-connector-python
Flask-WTF